Day 29-30: Work Session 4 (6-7:30)
During this work session I got stuck on #5 of the Project Euler problems. I decide to take the 75 minutes to figure out exactly what the code needs to do. I know that the code needs to loop over all values, starting at 2. If the number is divisible, it should go to the next number until it is no longer divisible. Then it should print the smallest divisible number.